# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EuroPulse AI — EU Macro Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardSummaryOverallRiskLevel = typeof DashboardSummaryOverallRiskLevel[keyof typeof DashboardSummaryOverallRiskLevel];


export const DashboardSummaryOverallRiskLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface DashboardSummary {
  inflationIndex: number;
  inflationChange: number;
  eurUsd: number;
  eurUsdChange: number;
  bondYield10y: number;
  bondYieldChange: number;
  overallRiskScore: number;
  overallRiskLevel?: DashboardSummaryOverallRiskLevel;
  riskScoreChange: number;
  lastUpdated?: string;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface Alert {
  id: string;
  countryCode: string;
  countryFlag: string;
  severity: AlertSeverity;
  headline: string;
  agentGenerated: boolean;
  timestamp: string;
}

export type HeatmapCountryRiskLevel = typeof HeatmapCountryRiskLevel[keyof typeof HeatmapCountryRiskLevel];


export const HeatmapCountryRiskLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface HeatmapCountry {
  countryCode: string;
  countryName: string;
  riskScore: number;
  riskLevel: HeatmapCountryRiskLevel;
  inflationRisk: number;
  fxStress: number;
  energyRisk: number;
  housingRisk: number;
  geopoliticalRisk: number;
}

export type NewsItemSeverity = typeof NewsItemSeverity[keyof typeof NewsItemSeverity];


export const NewsItemSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type TransmissionStepDirection = typeof TransmissionStepDirection[keyof typeof TransmissionStepDirection];


export const TransmissionStepDirection = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export interface TransmissionStep {
  label: string;
  detail: string;
  direction: TransmissionStepDirection;
}

export interface NewsItem {
  id: string;
  headline: string;
  severity: NewsItemSeverity;
  countryCode: string;
  countryFlag: string;
  countryName: string;
  /** @nullable */
  sector?: string | null;
  marketIndex: string;
  marketChange: number;
  whatHappened: string;
  whyItMatters: string;
  marketImpact: string;
  timestamp: string;
  /** @nullable */
  transmissionFlow?: TransmissionStep[] | null;
}

export type CountryRiskRiskLevel = typeof CountryRiskRiskLevel[keyof typeof CountryRiskRiskLevel];


export const CountryRiskRiskLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export type CountryRiskTrend = typeof CountryRiskTrend[keyof typeof CountryRiskTrend];


export const CountryRiskTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface CountryRisk {
  rank: number;
  countryCode: string;
  countryFlag: string;
  countryName: string;
  overallScore: number;
  riskLevel: CountryRiskRiskLevel;
  trend: CountryRiskTrend;
  inflationRisk: number;
  fxStress: number;
  energyRisk: number;
  housingRisk: number;
  geopoliticalRisk: number;
}

export type ForecastPredictionDirection = typeof ForecastPredictionDirection[keyof typeof ForecastPredictionDirection];


export const ForecastPredictionDirection = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface ChartDataPoint {
  date: string;
  /** @nullable */
  actual: number | null;
  /** @nullable */
  forecast?: number | null;
  /** @nullable */
  forecastHigh?: number | null;
  /** @nullable */
  forecastLow?: number | null;
}

export interface ForecastPrediction {
  id: string;
  metric: string;
  metricLabel: string;
  currentValue: number;
  unit: string;
  confidence: number;
  direction: ForecastPredictionDirection;
  keyDrivers: string[];
  chartData: ChartDataPoint[];
}

export interface Scenario {
  id: string;
  name: string;
  description: string;
  icon: string;
  defaultIntensity: number;
  minIntensity: number;
  maxIntensity: number;
  intensityUnit: string;
}

export interface SimulationInput {
  scenarioId: string;
  intensity: number;
}

export type SimTransmissionStepSeverity = typeof SimTransmissionStepSeverity[keyof typeof SimTransmissionStepSeverity];


export const SimTransmissionStepSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface SimTransmissionStep {
  step: number;
  label: string;
  icon: string;
  severity: SimTransmissionStepSeverity;
  detail: string;
}

export type SimImpactDirection = typeof SimImpactDirection[keyof typeof SimImpactDirection];


export const SimImpactDirection = {
  up: 'up',
  down: 'down',
} as const;

export type SimImpactImpactLevel = typeof SimImpactImpactLevel[keyof typeof SimImpactImpactLevel];


export const SimImpactImpactLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface SimImpact {
  metric: string;
  value: number;
  unit: string;
  direction: SimImpactDirection;
  impactLevel: SimImpactImpactLevel;
  description: string;
}

export interface SimulationResult {
  scenarioId: string;
  scenarioName: string;
  intensity: number;
  transmissionFlow: SimTransmissionStep[];
  impacts: SimImpact[];
  aiSummary: string;
  keyActions: string[];
  recommendedAction: string;
}

export type SidebarInsightAlertLevel = typeof SidebarInsightAlertLevel[keyof typeof SidebarInsightAlertLevel];


export const SidebarInsightAlertLevel = {
  green: 'green',
  yellow: 'yellow',
  orange: 'orange',
  red: 'red',
} as const;

export type SidebarInsightViewMode = typeof SidebarInsightViewMode[keyof typeof SidebarInsightViewMode];


export const SidebarInsightViewMode = {
  Simple: 'Simple',
  Banker: 'Banker',
  Trader: 'Trader',
} as const;

export interface ImpactCard {
  title: string;
  body: string;
}

export interface SidebarInsight {
  page: string;
  alertLevel: SidebarInsightAlertLevel;
  topInsight: string;
  viewMode?: SidebarInsightViewMode;
  impactCards: ImpactCard[];
  actions: string[];
  earlyWarning: string;
  recommendedAction: string;
}

export type GetNewsParams = {
country?: string;
severity?: GetNewsSeverity;
sector?: string;
date?: string;
};

export type GetNewsSeverity = typeof GetNewsSeverity[keyof typeof GetNewsSeverity];


export const GetNewsSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type GetRiskCountriesParams = {
riskType?: GetRiskCountriesRiskType;
};

export type GetRiskCountriesRiskType = typeof GetRiskCountriesRiskType[keyof typeof GetRiskCountriesRiskType];


export const GetRiskCountriesRiskType = {
  inflation: 'inflation',
  fx: 'fx',
  energy: 'energy',
  housing: 'housing',
  geopolitical: 'geopolitical',
  overall: 'overall',
} as const;

export type GetRiskHeatmapParams = {
riskType?: string;
};

export type GetSidebarInsightParams = {
page?: GetSidebarInsightPage;
};

export type GetSidebarInsightPage = typeof GetSidebarInsightPage[keyof typeof GetSidebarInsightPage];


export const GetSidebarInsightPage = {
  dashboard: 'dashboard',
  news: 'news',
  risk: 'risk',
  forecast: 'forecast',
  simulator: 'simulator',
} as const;