# lib/api-zod/src/generated/types/countryRisk.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EuroPulse AI — EU Macro Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CountryRiskRiskLevel } from './countryRiskRiskLevel';
import type { CountryRiskTrend } from './countryRiskTrend';

export interface CountryRisk {
  rank: number;
  countryCode: string;
  countryFlag: string;
  countryName: string;
  overallScore: number;
  riskLevel: CountryRiskRiskLevel;
  trend: CountryRiskTrend;
  inflationRisk: number;
  fxStress: number;
  energyRisk: number;
  housingRisk: number;
  geopoliticalRisk: number;
}