# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EuroPulse AI — EU Macro Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryOverallRiskLevel } from './dashboardSummaryOverallRiskLevel';

export interface DashboardSummary {
  inflationIndex: number;
  inflationChange: number;
  eurUsd: number;
  eurUsdChange: number;
  bondYield10y: number;
  bondYieldChange: number;
  overallRiskScore: number;
  overallRiskLevel?: DashboardSummaryOverallRiskLevel;
  riskScoreChange: number;
  lastUpdated?: string;
}