# lib/api-zod/src/generated/types/simulationInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EuroPulse AI — EU Macro Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationInput {
  scenarioId: string;
  intensity: number;
}