# lib/api-zod/src/generated/types/transmissionStep.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EuroPulse AI — EU Macro Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { TransmissionStepDirection } from './transmissionStepDirection';

export interface TransmissionStep {
  label: string;
  detail: string;
  direction: TransmissionStepDirection;
}